// src/generated/models/Stf_strategypocsModel.ts
export const Stf_strategypocspm_startmonth = {
  1: 'January',
  2: 'February',
  3: 'March',
  4: 'April',
  5: 'May',
  6: 'June',
  7: 'July',
  8: 'August',
  9: 'September',
  10: 'October',
  11: 'November',
  12: 'December'
} as const;
export type Stf_strategypocspm_startmonth = keyof typeof Stf_strategypocspm_startmonth;
export const Stf_strategypocsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Stf_strategypocsstatecode = keyof typeof Stf_strategypocsstatecode;
export const Stf_strategypocsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Stf_strategypocsstatuscode = keyof typeof Stf_strategypocsstatuscode;
export const Stf_strategypocsstf_pocstatus = {
  1: 'Active',
  2: 'Succeeded',
  3: 'Failed',
  4: 'Retired'
} as const;
export type Stf_strategypocsstf_pocstatus = keyof typeof Stf_strategypocsstf_pocstatus;
export const Stf_strategypocsstf_serviceexecutionmode = {
  1: 'TMS Task',
  2: 'Project'
} as const;
export type Stf_strategypocsstf_serviceexecutionmode = keyof typeof Stf_strategypocsstf_serviceexecutionmode;
export const Stf_strategypocsstf_strategytype = {
  989230000: 'Department Strategy',
  989230001: 'Market Strategy',
  989230002: 'Specialty Strategy',
  989230003: 'Service Strategy'
} as const;
export type Stf_strategypocsstf_strategytype = keyof typeof Stf_strategypocsstf_strategytype;

export interface Stf_strategypocsBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  "pm_Model@odata.bind"?: string;
  pm_startmonth?: Stf_strategypocspm_startmonth;
  statecode: Stf_strategypocsstatecode;
  statuscode?: Stf_strategypocsstatuscode;
  "stf_ConvertedTactic@odata.bind"?: string;
  stf_currentbaseline?: number;
  stf_experimentscope?: string;
  stf_from?: string;
  stf_killcondition?: string;
  stf_killconditiondate?: string;
  "stf_KPI@odata.bind"?: string;
  stf_kpitargetvalue?: number;
  stf_neededbudget?: number;
  "stf_POCCategory@odata.bind"?: string;
  stf_pocdescription?: string;
  stf_pocid?: string;
  stf_pocname?: string;
  stf_pocstatus?: Stf_strategypocsstf_pocstatus;
  "stf_Project@odata.bind"?: string;
  "stf_Region@odata.bind"?: string;
  stf_serviceexecutionmode?: Stf_strategypocsstf_serviceexecutionmode;
  "stf_Specialty@odata.bind"?: string;
  "stf_StrategyKPI@odata.bind"?: string;
  stf_strategypocid: string;
  stf_successcriteria?: string;
  stf_successduedate?: string;
  stf_target?: number;
  stf_to?: string;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}